### 2020-07-01 #2417 @whedon
[PRE REVIEW]: Extracting, Computing and Exploring the Parameters of Statistical Models using R
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/easystats/parameters" target ="_blank">https://github.com/easystats/parameters</a>
**Version:** 0.8.1
