### 2020-07-01 9445094
truncating overlengthy lines
