### 2020-07-01 #2417 @strengejacke
(truncated lines that exceeded page margins)
