### 2020-07-02 424be98
move (and re-export) value formatting functions
