### 2020-07-05 061b047
Allow 'threshold' to be passed to `efa_to_cfa()` when the model is fro `factor_analysis()`
