+++ NEWS.md
@@ -7,6 +7,7 @@
 
 ## Bug fixes
 
+* Allow 'threshold' to be passed to `efa_to_cfa()` when the model is fro `factor_analysis()`.
 * Allow correlation matrix to be passed to `factor_analysis()`.
 * Fix CRAN check issues.
 * Fix issue in `model_parameters()` for models with non-estimable parameters or statistics.
