+++ R/dof.R
@@ -1,6 +1,6 @@
 #' Degrees of Freedom (DoF)
 #'
-#' Estimate or extract degrees of freedom of models.
+#' Estimate or extract degrees of freedom of models parameters.
 #'
 #' @param model A statistical model.
 #' @param method Can be \code{"analytical"} (default, DoFs are estimated based on the model type), \code{"fit"}, in which case they are directly taken from the model if available (for Bayesian models, the goal (looking for help to make it happen) would be to refit the model as a frequentist one before extracting the DoFs), \code{"ml1"} (see \code{\link{dof_ml1}}), \code{"betwithin"} (see \code{\link{dof_betwithin}}), \code{"satterthwaite"} (see \code{\link{dof_satterthwaite}}), \code{"kenward"} (see \code{\link{dof_kenward}}) or \code{"any"}, which tries to extract DoF by any of those methods, whichever succeeds.
@@ -8,16 +8,17 @@
 #'
 #' @details Methods for calculating degrees of freedom:
 #' \itemize{
-#' \item \code{"analytical"} For models of class \code{lmerMod}, Kenward-Roger approximated degrees of freedoms are calculated, for other models, \code{n-k} (number of observations minus number of parameters).
-#' \item \code{"fit"} Tries to extract residual degrees of freedom, and returns \code{Inf} if residual degrees of freedom could not be extracted.
-#' \item \code{"any"} First tries to extract residual degrees of freedom, and if these are not available, extracts analytical degrees of freedom.
-#' \item \code{"nokr"} Same as \code{"analytical"}, but does not Kenward-Roger approximation for models of class \code{lmerMod}. Instead, always uses \code{n-k} to calculate df for any model.
-#' \item \code{"wald"} Returns \code{Inf}.
-#' \item \code{"kenward"} Calls \code{\link{dof_kenward}}.
-#' \item \code{"satterthwaite"} Calls \code{\link{dof_satterthwaite}}.
-#' \item \code{"ml1"} Calls \code{\link{dof_ml1}}.
-#' \item \code{"betwithin"} Calls \code{\link{dof_betwithin}}.
+#' \item \code{"analytical"} for models of class \code{lmerMod}, Kenward-Roger approximated degrees of freedoms are calculated, for other models, \code{n-k} (number of observations minus number of parameters).
+#' \item \code{"fit"} tries to extract residual degrees of freedom, and returns \code{Inf} if residual degrees of freedom could not be extracted.
+#' \item \code{"any"} first tries to extract residual degrees of freedom, and if these are not available, extracts analytical degrees of freedom.
+#' \item \code{"nokr"} same as \code{"analytical"}, but does not Kenward-Roger approximation for models of class \code{lmerMod}. Instead, always uses \code{n-k} to calculate df for any model.
+#' \item \code{"wald"} returns \code{Inf}.
+#' \item \code{"kenward"} calls \code{\link{dof_kenward}}.
+#' \item \code{"satterthwaite"} calls \code{\link{dof_satterthwaite}}.
+#' \item \code{"ml1"} calls \code{\link{dof_ml1}}.
+#' \item \code{"betwithin"} calls \code{\link{dof_betwithin}}.
 #' }
+#' For models with z-statistic, the returned degrees of freedom for model parameters is \code{Inf} (unless \code{method = "ml1"} or \code{method = "betwithin"}), because there is only one distribution for the related test statistic.
 #'
 #' @examples
 #' model <- lm(Sepal.Length ~ Petal.Length * Species, data = iris)
@@ -47,6 +48,7 @@ degrees_of_freedom <- function(model, ...) {
 }
 
 
[…]
