+++ NAMESPACE
@@ -718,6 +718,7 @@ importFrom(insight,find_predictors)
 importFrom(insight,find_random)
 importFrom(insight,find_random_slopes)
 importFrom(insight,find_response)
+importFrom(insight,find_statistic)
 importFrom(insight,find_terms)
 importFrom(insight,format_bf)
 importFrom(insight,format_ci)
