+++ R/equivalence_test.R
@@ -36,8 +36,11 @@ bayestestR::equivalence_test
 #' speaking, the idea behind equivalence testing in a frequentist framework is
 #' to check whether an estimate and its uncertainty (i.e. confidence interval)
 #' falls within a region of "practical equivalence". Depending on the rule for
-#' this test (see below), statistical significance may not be used to reject the
-#' null hypothesis.
+#' this test (see below), statistical significance does not necessarily indicate
+#' whether the null hypothesis can be rejected or not, i.e. the classical
+#' interpretation of the p-value may differ from the results returned from
+#' the equivalence test.
+#'
 #' \subsection{Calculation of equivalence testing}{
 #'   \describe{
 #'     \item{"bayes" - Bayesian rule (Kruschke 2018)}{
@@ -70,12 +73,21 @@ bayestestR::equivalence_test
 #'   }
 #' }
 #' \subsection{Levels of Confidence Intervals used for Equivalence Testing}{
-#'   For \code{rule = "cet"}, "narrow" confidence intervals are used for
+#'   For \code{rule = "classic"}, "narrow" confidence intervals are used for
 #'   equivalence testing. "Narrow" means, the the intervals is not 1 - alpha,
 #'   but 1 - 2 * alpha. Thus, if \code{ci = .95}, alpha is assumed to be 0.05
-#'   and internally a ci-level of 0.90 is used. \code{rule = "classic"} uses
+#'   and internally a ci-level of 0.90 is used. \code{rule = "cet"} uses
 #'   both regular and narrow confidence intervals, while \code{rule = "bayes"}
 #'   only uses the regular intervals.
+#' }
+#' \subsection{Second Generation p-Value (SGPV)}{
+#'   Second generation p-values (SGPV) were proposed as a statistic
+#'   that represents \dQuote{the proportion of data-supported hypotheses
+#'   that are also null hypotheses} \cite{(Blume et al. 2018)}. This statistic
+#'   is actually computed in the same way as the percentage inside the ROPE as
+#'   returned by \code{equivalence_test()} (see \cite{Lakens and Delacre 2020}
+#'   for details on computation of the SGPV). Thus, the \code{"inside ROPE"}
+#'   column reflects the SGPV.
 #' }
 #' \subsection{Adjustment for multiple testing}{
 #'   The calculation of p-values is somewhat "experimental". For parameters, where H0...
@@ -96,10 +108,12 @@ bayestestR::equivalence_test
 #'
 #' @references
 #' \itemize{
+#'   \item Blume, J. D., D'Agostino McGowan, L., Dupont, W. D., & Greevy, R. A. (2018). Second-generation p-values: Improved rigor, reproducibility, & transparency in statistical analyses. PLOS ONE, 13(3), e0188299. https://doi.org/10.1371/journal.pone.0188299
 #'   \item Campbell, H., & Gustafson, P. (2018). Conditional equivalence testing: An alternative remedy for publication bias. PLOS ONE, 13(4), e0195145. doi: 10.1371/journal.pone.0195145
 #'   \item Kruschke, J. K. (2014). Doing Bayesian data analysis: A tutorial with R, JAGS, and Stan. Academic Press
 #'   \item Kruschke, J. K. (2018). Rejecting or accepting parameter values in Bayesian estimation. Advances in Methods and Practices in Psychological Science, 1(2), 270-280. doi: 10.1177/2515245918771304
 #'   \item Lakens, D. (2017). Equivalence Tests: A Practical Primer for t Tests, Correlations, and Meta-Analyses. Social Psychological and Personality Science, 8(4), 355–362. doi: 10.1177/1948550617697177
[…]
