+++ R/dof.R
@@ -237,11 +237,12 @@ degrees_of_freedom.betamfx <- degrees_of_freedom.logitor
   if (is.null(method)) {
     return(TRUE)
   }
+  if (inherits(model, c("polr", "glm")) && method %in% c("profile", "wald")) {
+    return(TRUE)
+  }
+
   info <- insight::model_info(model, verbose = FALSE)
   if (is.null(info) || !info$is_mixed) {
-    if (method %in% c("profile", "wald")) {
-      return(TRUE)
-    }
     return(FALSE)
   }
   method <- tolower(method)
