### 2020-07-18 af1bad1
include docs for GLM, make df_method more explicit
