+++ man/model_parameters.default.Rd
@@ -6,6 +6,7 @@
 \alias{model_parameters.poissonmfx}
 \alias{model_parameters.betamfx}
 \alias{model_parameters.default}
+\alias{model_parameters.glm}
 \alias{model_parameters.betareg}
 \alias{model_parameters.clm2}
 \alias{model_parameters.glmx}
@@ -52,6 +53,19 @@
 \method{model_parameters}{default}(
   model,
   ci = 0.95,
+  bootstrap = FALSE,
+  iterations = 1000,
+  standardize = NULL,
+  exponentiate = FALSE,
+  robust = FALSE,
+  p_adjust = NULL,
+  ...
+)
+
+\method{model_parameters}{glm}(
+  model,
+  ci = 0.95,
+  df_method = "profile",
   bootstrap = FALSE,
   iterations = 1000,
   standardize = NULL,
@@ -117,6 +131,8 @@
 \item{...}{Arguments passed to or from other methods. For instance, when \code{bootstrap = TRUE}, arguments like \code{ci_method} are passed down to \code{\link[bayestestR]{describe_posterior}}.}
 
 \item{component}{Model component for which parameters should be shown. May be one of \code{"conditional"}, \code{"precision"} (\pkg{betareg}), \code{"scale"} (\pkg{ordinal}), \code{"extra"} (\pkg{glmx}), \code{"marginal"} (\pkg{mfx}) or \code{"all"}.}
+
+\item{df_method}{Method for computing degrees of freedom for confidence intervals (CI). Only applies to models of class \code{glm} or \code{polr}. May be \code{"profile"} or \code{"wald"}.}
 }
 \value{
 A data frame of indices related to the model's parameters.
