+++ NAMESPACE
@@ -192,6 +192,7 @@ S3method(model_parameters,gam)
 S3method(model_parameters,gamlss)
 S3method(model_parameters,gamm)
 S3method(model_parameters,glht)
+S3method(model_parameters,glm)
 S3method(model_parameters,glmmTMB)
 S3method(model_parameters,glmx)
 S3method(model_parameters,htest)
@@ -215,6 +216,7 @@ S3method(model_parameters,parameters_efa)
 S3method(model_parameters,parameters_pca)
 S3method(model_parameters,poissonirr)
 S3method(model_parameters,poissonmfx)
+S3method(model_parameters,polr)
 S3method(model_parameters,principal)
 S3method(model_parameters,probitmfx)
 S3method(model_parameters,rlmerMod)
