+++ NAMESPACE
@@ -57,6 +57,7 @@ S3method(ci,gee)
 S3method(ci,geeglm)
 S3method(ci,glht)
 S3method(ci,glm)
+S3method(ci,glmm)
 S3method(ci,glmmTMB)
 S3method(ci,glmmadmb)
 S3method(ci,glmrob)
@@ -150,6 +151,7 @@ S3method(equivalence_test,wbm)
 S3method(equivalence_test,zeroinfl)
 S3method(factor_analysis,data.frame)
 S3method(format_parameters,default)
+S3method(format_parameters,glmm)
 S3method(format_parameters,parameters_model)
 S3method(format_parameters,rma)
 S3method(kurtosis,data.frame)
@@ -193,6 +195,7 @@ S3method(model_parameters,gamlss)
 S3method(model_parameters,gamm)
 S3method(model_parameters,glht)
 S3method(model_parameters,glm)
+S3method(model_parameters,glmm)
 S3method(model_parameters,glmmTMB)
 S3method(model_parameters,glmx)
 S3method(model_parameters,htest)
@@ -307,6 +310,7 @@ S3method(p_value,gee)
 S3method(p_value,geeglm)
 S3method(p_value,glht)
 S3method(p_value,glimML)
+S3method(p_value,glmm)
 S3method(p_value,glmmTMB)
 S3method(p_value,glmx)
 S3method(p_value,gls)
@@ -559,6 +563,7 @@ S3method(standard_error,geeglm)
 S3method(standard_error,glht)
 S3method(standard_error,glimML)
 S3method(standard_error,glm)
+S3method(standard_error,glmm)
 S3method(standard_error,glmmTMB)
 S3method(standard_error,glmx)
 S3method(standard_error,gls)
