### 2020-07-19 easystats/parameters#274 @cmaimone
Paper: include comparison to similar packages
> State of the field: Do the authors describe how this software compares to other commonly-used packages?

The discussion of how parameters relates to broom in the readme covers this, but it is not included in the paper.

It also may be useful to include reference to the set of packages dedicated to reporting of model results (such as stargazer and finalfit), since that is one of the aims of parameters called out in the paper.  Not necessarily a comparison to each of them, but how the general functionality of parameters and what parameters aims to do is similar to or different from these types of packages.

Re: https://github.com/openjournals/joss-reviews/issues/2445
