### 2020-07-19 easystats/parameters#276 @cmaimone
Vignette parameters_selection: Mixed and Bayesian models section clarification 
This vignette section https://easystats.github.io/parameters/articles/parameters_selection.html#mixed-and-bayesian-models is a bit confusing.  By running the code in the example with and without select_parameters(), it appears that when select_parameters() is used with lmer (at least with the example model), it simply has no effect.  This was not clear from the text in the section.  I'm guessing that "However, this procedures is not available for other types of models, such as mixed or Bayesian models." was meant to convey that, but I thought it was talking about a particular selection algorithm, not the entire function.

Is the effect the same for Bayesian models (not included in the vignette).  The documentation suggests it wouldn't be: "For Bayesian models, it uses the projpred package."

If select_parameters() is intended to simply have no effect for mixed (or other) models, separate from the vignette, it might also be nice to issue a warning to the user in such cases.
