### 2020-07-19 easystats/parameters#277 @cmaimone
Paper: consider broader set of citations
The JOSS paper includes citations for the packages directly imported by parameters, but you may consider also citing packages that are used directly by the functions in parameters.  For example, when running the vignette code, I encountered (very helpful!) messages to install cAIC4 and sandwich.  These packages were needed not just to run the examples of different models (like lme4 is), but rather to use functions in parameters.  There are probably other packages like this too.

I very much appreciate that you did not make all of these packages dependencies of parameters.  However, since you rely on them, citations at least may be warranted.

Re: https://github.com/openjournals/joss-reviews/issues/2445
