+++ R/select_parameters.R
@@ -1,9 +1,25 @@
 #' Automated selection of model parameters
 #'
-#' This function performs an automated selection of the 'best' parameters, updating and returning the "best" model. For frequentist simple GLMs, it performs an AIC-based stepwise selection. For Bayesian models, it uses the \code{projpred} package.
+#' This function performs an automated selection of the 'best' parameters, updating and returning the "best" model.
 #'
-#' @param model A statistical model.
+#' @param model A statistical model (of class \code{lm}, \code{glm}, \code{merMod}, \code{stanreg} or \code{brmsfit}).
 #' @param ... Arguments passed to or from other methods.
+#'
+#' @details
+#'   \subsection{Classical lm and glm}{
+#'     For frequentist GLMs, \code{select_parameters()} performs an AIC-based
+#'     stepwise selection.
+#'   }
+#'   \subsection{Mixed models}{
+#'     For mixed models of class \code{merMod}, stepwise selection is
+#'     based on \code{\link[cAIC4:stepcAIC]{stepcAIC()}}. This step function
+#'     only searches the "best" model based on the random effects structure,
+#'     i.e. \code{select_parameters()} adds or excludes random effects until
+#'     the cAIC can't be improved further.
+#'   }
+#'   \subsection{Bayesian models}{
+#'     For Bayesian models, it uses the \code{projpred} package.
+#'   }
 #'
 #' @examples
 #' model <- lm(mpg ~ ., data = mtcars)
@@ -66,7 +82,7 @@ select_parameters.lm <- function(model, direction = "both", steps = 1000, k = 2,
 
 
 
-
+#' @importFrom insight find_random
 #' @rdname select_parameters
 #' @export
 select_parameters.merMod <- function(model, direction = "backward", steps = 1000, ...) {
@@ -87,13 +103,20 @@ select_parameters.merMod <- function(model, direction = "backward", steps = 1000
   #   nums <- NULL
   # }
 
-  best <- cAIC4::stepcAIC(model,
-    # groupCandidates = factors,
+  factors <- unique(c(
+    insight::find_random(model, split_nested = FALSE, flatten = TRUE),
+    insight::find_random(model, split_nested = TRUE, flatten = TRUE)
+  ))
+  factors <- gsub(":", "/", factors, fixed = TRUE)
+
+
+  best <- suppressMessages(suppressWarnings(cAIC4::stepcAIC(model,
     # slopeCandidates = nums,
+    groupCandidates = factors,
     direction = direction,
     steps = steps,
     allowUseAcross = TRUE
-  )$finalModel
+  )$finalModel))
 
 
   # Using MuMIn's dredge(): works nicely BUT throws unnecessary warnings and requires to set global options for na.action even tho no NaNs.
