+++ NAMESPACE
@@ -429,6 +429,7 @@ S3method(reshape_loadings,parameters_efa)
 S3method(se_satterthwaite,default)
 S3method(se_satterthwaite,gls)
 S3method(se_satterthwaite,lme)
+S3method(select_parameters,brmsfit)
 S3method(select_parameters,lm)
 S3method(select_parameters,merMod)
 S3method(select_parameters,stanreg)
