### 2020-07-19 easystats/parameters#276 @strengejacke
Thanks! We revised the vignette, which now explains more in detail how `select_parameters()` works for mixed models, slightly revised the examples, and added an example for Bayesian models to the vignette. It should now be more clear how `select_parameters()` works for mixed and Bayesian models.

Furthermore, we revised the docs of the `select_parameters()` function, so the help-file is more clear to users now.
