### 2020-07-19 easystats/parameters#276 @strengejacke
(pinging https://github.com/openjournals/joss-reviews/issues/2445)
