+++ paper/paper.bib
@@ -1,12 +1,3 @@
-@Manual{ludecke2019see,
-    title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
-    author = {Daniel Lüdecke and Philip Waggoner and Mattan S. Ben-Shachar and Dominique Makowski},
-    year = {2019},
-    note = {R package version 0.4.0},
-    url = {https://easystats.github.io/see/}
-  }
-
-
 @Article{makowski2019bayetestR,
 	doi = {10.21105/joss.01541},
 	url = {https://doi.org/10.21105/joss.01541},
@@ -34,12 +25,12 @@
 
 
 @Manual{rcore,
-  title = {R: A Language and Environment for Statistical Computing},
+  title = {{R}: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2020},
-  url = {https://www.R-project.org/},
+  url = {https://www.R-project.org/}
 }
 
 
@@ -51,7 +42,7 @@
   volume = {16},
   number = {9},
   pages = {1--16},
-  doi = {10.18637/jss.v016.i09},
+  doi = {10.18637/jss.v016.i09}
 }
 
 
@@ -61,5 +52,92 @@ Corrections},
   author = {James Pustejovsky},
   year = {2020},
   note = {R package version 0.4.2},
-  url = {https://CRAN.R-project.org/package=clubSandwich},
+  url = {https://CRAN.R-project.org/package=clubSandwich}
+}
+
+
+
+@article{lakens2020equivalence,
+	title = {Equivalence Testing for Psychological Research: A Tutorial},
+	volume = {1},
+	issn = {2515-2459, 2515-2467},
+	url = {http://journals.sagepub.com/doi/10.1177/2515245918770963},
+	doi = {10.1177/2515245918770963},
+	language = {en},
+	number = {2},
+	journal = {Advances in Methods and Practices in Psychological Science},
+	author = {Lakens, Daniël and Scheel, Anne M. and Isager, Peder M.},
+	month = jun,
+	year = {2018},
+	pages = {259--269}
+}
+
+
+@Manual{luedecke2020performance,
+  title = {performance: Assessment of Regression Models Performance},
+  author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner and Indrajeet Patil},
+  year = {2020},
+  note = {R package version 0.4.7.1},
+  url = {https://easystats.github.io/performance/}
+}
+
+
+@Manual{robinson_broom_2020,
+  title = {broom: Convert Statistical Objects into Tidy Tibbles},
+  author = {David Robinson and Alex Hayes and Simon Couch},
+  year = {2020},
+  note = {R package version 0.7.0},
+  url = {https://CRAN.R-project.org/package=broom}
+}
+
+
+@Manual{hlavac_stargazer_2018,
+  title = {stargazer: Well-Formatted Regression and Summary Statistics Tables},
+  author = {Marek Hlavac},
+  year = {2018},
+  note = {R package version 5.2.2},
+  organization = {Central European Labour Studies Institute (CELSI)},
+  address = {Bratislava, Slovakia},
+  url = {https://CRAN.R-project.org/package=stargazer},
+}
+
+
+@Manual{harrison2020finalfit,
+  title = {finalfit: Quickly Create Elegant Regression Results Tables and Plots when
+Modelling},
+  author = {Ewen Harrison and Tom Drake and Riinu Ots},
+  year = {2020},
+  note = {R package version 1.0.2},
+  url = {https://CRAN.R-project.org/package=finalfit},
+}
+
+
[…]
