### 2020-07-20 easystats/parameters#274 @strengejacke
Thanks for this comment! We have added a section to the paper that includes a comparison of similarities and differences of the _parameters_ package to other packages with comparable functionality. The new section in the paper now reads:

# Comparison to other Packages

**parameters** functionality is in parts comparable to packages like **broom** (Robinson, Hayes, & Couch, 2020), **finalfit** (Harrison, Drake, & Ots, 2020) or **stargazer** (Hlavac, 2018) (and maybe some more). Yet, there are some notable differences, e.g.:

- **broom** (via `glance()`), **finalfit** (via `ff_metrics()`) and **stargazer** (via `stargazer()`) report fit indices such as R2 or AIC by default, **parameters** does not. However, there is a dedicated package in the **easystats** project for assessing regression model quality and fit indices, **performance** (Lüdecke et al., 2020b).
- **parameters** easily allows to compute standardized estimates, robust estimation, small-sample-size corrections for degrees of freedom (like *Satterthwaite* or *Kenward-Roger*), bootstrapping or simulating parameters, and feature reduction. Furthermore, **parameters** provides functions to test for the presence or absence of an effect (_equivalence testing_, see Lakens, Scheel, & Isager, 2018).
- For most functions, [easy-to-use `plot()`-methods](https://easystats.github.io/see/articles/parameters.html) exist to quickly create nice looking plots (powered by the **see** package (Lüdecke et al., 2020a)).
- **parameters** is a very lightweight package. Its main functionality only relies on the **insight** and the **bayestestR** packages (Lüdecke, Waggoner, & Makowski, 2019; Makowski, Ben-Shachar, & Lüdecke, 2019) to access and process information contained in models, and these packages in turn only depend on R core packages. However, additional features that do not belong to the core functions of **parameters** require the installation of other packages, such as **sandwich** (Zeileis, 2006) for robust estimation, **psych** (Revelle, 2019) for factor analysis or PCA or **cAIC4** (Saefken, Ruegamer, Kneib, & Greven, 2018) for parameter selection for mixed models.
