### 2020-07-20 easystats/parameters#277 @strengejacke
In the section "Comparison to other Packages" (see #274), we included a bullet-point about the necessary and optional dependencies of the package, and included references to other packages that are only required for additional functionality. See the new paragraph in the paper in my comment in #274.
