### 2020-07-20 #2445 @strengejacke
@cmaimone Dear Christina, thank you for your reviews regarding our package (website) and the paper. We have addressed your issues and included a response to your requests in the related GitHub issues (https://github.com/easystats/parameters/issues/274, https://github.com/easystats/parameters/issues/276 and https://github.com/easystats/parameters/issues/277).

We hope that our revisions adequately address your proposed changes. Please let us know if there are any further things you like us to revise.
