+++ paper/paper.bib
@@ -141,3 +141,12 @@ and Color Palettes for 'ggplot2'},
   eprint = {1803.05664},
   month = {march}
 }
+
+@Article{benshachar2020effecsize,
+  title = {Compute and interpret indices of effect size},
+  author = {Mattan S. Ben-Shachar and Dominique Makowski and Daniel Lüdecke},
+  journal = {CRAN},
+  year = {2020},
+  note = {R package},
+  url = {https://easystats.github.io/effectsize},
+}
