### 2020-07-20 f328356
fix code in vignette
