+++ tests/testthat/test-gee.R
@@ -8,7 +8,7 @@ if (require("testthat") &&
     expect_equal(
       ci(m1)$CI_low,
       c(30.90044, -17.76184, -22.48406),
-      tolerance = 1e-4
+      tolerance = 1e-3
     )
   })
 
@@ -16,15 +16,15 @@ if (require("testthat") &&
     expect_equal(
       standard_error(m1)$SE,
       c(2.80028, 3.96019, 3.96019),
-      tolerance = 1e-4
+      tolerance = 1e-3
     )
   })
 
   test_that("p_value", {
     expect_equal(
       p_value(m1)$p,
-      c(0, 0.014717, 0.000501),
-      tolerance = 1e-4
+      c(0, 0.01157, 2e-04),
+      tolerance = 1e-3
     )
   })
 
@@ -32,7 +32,7 @@ if (require("testthat") &&
     expect_equal(
       model_parameters(m1)$Coefficient,
       c(36.38889, -10, -14.72222),
-      tolerance = 1e-4
+      tolerance = 1e-3
     )
   })
 }
