+++ paper/paper.bib
@@ -112,13 +112,14 @@ Modelling},
 }
 
 
-@Manual{ludecke2020see,
-  title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes
-and Color Palettes for 'ggplot2'},
-  author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner and Mattan S. Ben-Shachar},
+@Article{ludecke2020see,
+  title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
+  author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Philip Waggoner and Dominique Makowski},
+  journal = {CRAN},
   year = {2020},
   note = {R package version 0.5.1},
-  url = {https://easystats.github.io/see/}
+  doi = {10.5281/zenodo.3952153},
+  url = {https://easystats.github.io/see}
 }
 
 
