+++ paper/paper.bib
@@ -73,12 +73,14 @@ Corrections},
 }
 
 
-@Manual{luedecke2020performance,
+@Article{luedecke2020performance,
   title = {performance: Assessment of Regression Models Performance},
   author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner and Indrajeet Patil},
+  journal = {CRAN},
   year = {2020},
-  note = {R package version 0.4.7.1},
-  url = {https://easystats.github.io/performance/}
+  note = {R package},
+  doi = {10.5281/zenodo.3952174},
+  url = {https://easystats.github.io/performance}
 }
 
 
