+++ paper/paper.bib
@@ -151,5 +151,6 @@ Modelling},
   journal = {CRAN},
   year = {2020},
   note = {R package},
+  doi = {10.5281/zenodo.3952214},
   url = {https://easystats.github.io/effectsize},
 }
