### 2020-07-20 #2445 @strengejacke
We created Zenodo-repositories to assign our packages with a DOI, so all references should have a DOI now (where possible).
