+++ paper/paper.bib
@@ -13,7 +13,7 @@
 
 
 @Article{ludecke2019insight,
-  title = {insight: A Unified Interface to Access Information from Model Objects in {R}},
+  title = {{insight}: A Unified Interface to Access Information from Model Objects in {R}},
   volume = {4},
   doi = {10.21105/joss.01412},
   number = {38},
@@ -47,7 +47,7 @@
 
 
 @Manual{pustejovsky2020,
-  title = {clubSandwich: Cluster-Robust (Sandwich) Variance Estimators with Small-Sample
+  title = {{clubSandwich}: Cluster-Robust (Sandwich) Variance Estimators with Small-Sample
 Corrections},
   author = {James Pustejovsky},
   year = {2020},
@@ -74,7 +74,7 @@ Corrections},
 
 
 @Article{luedecke2020performance,
-  title = {performance: Assessment of Regression Models Performance},
+  title = {{performance}: Assessment of Regression Models Performance},
   author = {Daniel Lüdecke and Dominique Makowski and Philip Waggoner and Indrajeet Patil},
   journal = {CRAN},
   year = {2020},
@@ -85,7 +85,7 @@ Corrections},
 
 
 @Manual{robinson_broom_2020,
-  title = {broom: Convert Statistical Objects into Tidy Tibbles},
+  title = {{broom}: Convert Statistical Objects into Tidy Tibbles},
   author = {David Robinson and Alex Hayes and Simon Couch},
   year = {2020},
   note = {R package version 0.7.0},
@@ -94,7 +94,7 @@ Corrections},
 
 
 @Manual{hlavac_stargazer_2018,
-  title = {stargazer: Well-Formatted Regression and Summary Statistics Tables},
+  title = {{stargazer}: Well-Formatted Regression and Summary Statistics Tables},
   author = {Marek Hlavac},
   year = {2018},
   note = {R package version 5.2.2},
@@ -105,7 +105,7 @@ Corrections},
 
 
 @Manual{harrison2020finalfit,
-  title = {finalfit: Quickly Create Elegant Regression Results Tables and Plots when
+  title = {{finalfit}: Quickly Create Elegant Regression Results Tables and Plots when
 Modelling},
   author = {Ewen Harrison and Tom Drake and Riinu Ots},
   year = {2020},
@@ -115,7 +115,7 @@ Modelling},
 
 
 @Article{ludecke2020see,
-  title = {see: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
+  title = {{see}: Visualisation Toolbox for 'easystats' and Extra Geoms, Themes and Color Palettes for 'ggplot2'},
   author = {Daniel Lüdecke and Mattan S. Ben-Shachar and Philip Waggoner and Dominique Makowski},
   journal = {CRAN},
   year = {2020},
@@ -126,7 +126,7 @@ Modelling},
 
 
 @Manual{revelle_psych_2019,
-  title = {psych: Procedures for Psychological, Psychometric, and Personality Research},
+  title = {{psych}: Procedures for Psychological, Psychometric, and Personality Research},
   author = {William Revelle},
   organization = {Northwestern University},
   address = {Evanston, Illinois},
@@ -137,7 +137,7 @@ Modelling},
 
 
 @Article{saefken_caic4_2018,
-  title = {Conditional Model Selection in Mixed-Effects Models with cAIC4},
+  title = {Conditional Model Selection in Mixed-Effects Models with {cAIC4}},
   author = {Benjamin Saefken and David Ruegamer and Thomas Kneib and Sonja Greven},
   journal = {ArXiv e-prints},
   year = {2018},
@@ -146,7 +146,7 @@ Modelling},
 }
 
[…]
