### 2020-07-21 ae590b7
also allow formula-notation.
