+++ DESCRIPTION
@@ -47,7 +47,7 @@ BugReports: https://github.com/easystats/parameters/issues
 Depends:
     R (>= 3.2)
 Imports:
-    bayestestR (>= 0.5.0),
+    bayestestR (>= 0.7.0),
     insight (>= 0.8.1),
     methods,
     stats,
