+++ R/standard_error.R
@@ -964,29 +964,15 @@ standard_error.cpglmm <- function(model, ...) {
 
 #' @export
 standard_error.rq <- function(model, ...) {
-  se <- tryCatch(
-    {
-      cs <- suppressWarnings(stats::coef(summary(model)))
-      se_column <- intersect(c("Std Error", "Std. Error"), colnames(cs))
-      if (length(se_column)) {
-        cs[, se_column]
-      } else {
-        vc <- insight::get_varcov(model)
-        as.vector(sqrt(diag(vc)))
-      }
-    },
-    error = function(e) {
-      vc <- insight::get_varcov(model)
-      as.vector(sqrt(diag(vc)))
-    }
-  )
+  se <- .get_quantreg_se(model)
+  if (is.null(se)) {
+    vc <- insight::get_varcov(model)
+    se <- as.vector(sqrt(diag(vc)))
+  }
 
   params <- insight::get_parameters(model)
-
-  .data_frame(
-    Parameter = params$Parameter,
-    SE = se
-  )
+  params$SE <- se
+  params[intersect(colnames(params), c("Parameter", "SE", "Component"))]
 }
 
 #' @export
@@ -994,8 +980,6 @@ standard_error.crq <- standard_error.rq
 
 #' @export
 standard_error.nlrq <- standard_error.rq
-
-
 
 #' @export
 standard_error.rqss <- function(model, component = c("all", "conditional", "smooth_terms"), ...) {
@@ -1577,6 +1561,46 @@ standard_error.blavaan <- function(model, ci = .95, ...) {
   }
 
   names(se) <- .remove_backticks_from_string(names(se))
+  se
+}
+
+
+#' @importFrom stats coef setNames
+#' @importFrom insight get_varcov
+.get_quantreg_se <- function(model) {
+  se <- tryCatch(
+    {
+      cs <- suppressWarnings(stats::coef(summary(model)))
+      se_column <- intersect(c("Std Error", "Std. Error"), colnames(cs))
+      if (length(se_column)) {
+        cs[, se_column]
+      } else {
+        vc <- insight::get_varcov(model)
+        as.vector(sqrt(diag(vc)))
+      }
+    },
+    error = function(e) { NULL }
+  )
+
+  if (is.null(se)) {
+    se <- tryCatch(
+      {
+        sc <- summary(model)
+        if (all(unlist(lapply(sc, is.list)))) {
+          list_sc <- lapply(sc, function(i) {
+            .x <- as.data.frame(i)
+            .x$Parameter <- rownames(.x)
+            .x
+          })
+          out <- do.call(rbind, list_sc)
+          se <- stats::setNames(out$coefficients.Std.Error, sprintf("tau (%g)", out$tau))
+        } else {
+          se <- stats::setNames(unname(sc$coefficients[, 4]), names(sc$coefficients[, 4]))
+        }
+      },
+      error = function(e) { NULL }
+    )
+  }
   se
 }
 
