+++ tests/testthat/test-model_parameters_robust.R
@@ -1,6 +1,7 @@
 if (require("testthat") &&
   require("parameters") &&
   require("sandwich") &&
+  require("clubSandwich") &&
   require("effectsize")) {
 
   data(mtcars)
