+++ man/standard_error_robust.Rd
@@ -43,7 +43,7 @@ covariance matrix estimators).}
 
 \item{vcov_type}{Character vector, specifying the estimation type for the
 robust covariance matrix estimation (see \code{\link[sandwich:vcovHC]{vcovHC()}}
-or \code{\link[clubSandwich:vcovCR]{vcovCR()}} for details).}
+or \code{clubSandwich::vcovCR()} for details).}
 
 \item{vcov_args}{List of named vectors, used as additional arguments that
 are passed down to the \pkg{sandwich}-function specified in \code{vcov_estimation}.}
