+++ tests/testthat/test-quantreg.R
@@ -33,19 +33,21 @@ if (require("testthat") &&
 
   # crq ---------
 
-  set.seed(123)
-  n <- 200
-  x <- rnorm(n)
-  y <- 5 + x + rnorm(n)
-  c <- 4 + x + rnorm(n)
-  d <- (y > c)
+  if (packageVersion("insight") > "0.9.0") {
+    set.seed(123)
+    n <- 200
+    x <- rnorm(n)
+    y <- 5 + x + rnorm(n)
+    c <- 4 + x + rnorm(n)
+    d <- (y > c)
 
-  dat <- data.frame(y, x, c, d)
-  m1 <- crq(survival::Surv(pmax(y,c), d, type = "left") ~ x, method = "Portnoy", data = dat)
+    dat <- data.frame(y, x, c, d)
+    m1 <- crq(survival::Surv(pmax(y,c), d, type = "left") ~ x, method = "Portnoy", data = dat)
 
-  mp <- model_parameters(m1)
-  test_that("mp_rq", {
-    expect_identical(mp$Parameter, c("(Intercept)", "x", "(Intercept)", "x", "(Intercept)", "x", "(Intercept)", "x"))
-    expect_equal(mp$Coefficient, c(4.26724, 0.97534, 4.84961, 0.92638, 5.21843, 0.98038, 5.91301, 0.97382), tolerance = 1e-3)
-  })
+    mp <- model_parameters(m1)
+    test_that("mp_rq", {
+      expect_identical(mp$Parameter, c("(Intercept)", "x", "(Intercept)", "x", "(Intercept)", "x", "(Intercept)", "x"))
+      expect_equal(mp$Coefficient, c(4.26724, 0.97534, 4.84961, 0.92638, 5.21843, 0.98038, 5.91301, 0.97382), tolerance = 1e-3)
+    })
+  }
 }
