### 2020-07-24 #2445 @strengejacke
@sbenthall @Ohlsen After revising the paper and package based on the first review, CRAN requested an update from the _parameters_ package. Thus, I needed to increase the version number to indicate a "stable" release, and the GitHub-version will be 0.8.2.x the next days, and no longer 0.8.1.x - so don't be confused, @Ohlsen , when you start your review.
