### 2020-07-24 688da23
Make sure clubSandwich is used conditionally in all places
