+++ man/standard_error_robust.Rd
@@ -68,15 +68,19 @@ These functions rely on the \pkg{sandwich} or \pkg{clubSandwich} package
   and will thus only work for those models supported by those packages.
 }
 \examples{
-# robust standard errors, calling sandwich::vcovHC(type="HC3") by default
-model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
-standard_error_robust(model)
+if (require("sandwich")) {
+  # robust standard errors, calling sandwich::vcovHC(type="HC3") by default
+  model <- lm(Petal.Length ~ Sepal.Length * Species, data = iris)
+  standard_error_robust(model)
+}
 
-# cluster-robust standard errors, using clubSandwich
-iris$cluster <- factor(rep(LETTERS[1:8], length.out = nrow(iris)))
-standard_error_robust(
-  model,
-  vcov_type = "CR2",
-  vcov_args = list(cluster = iris$cluster)
-)
+if (require("clubSandwich")) {
+  # cluster-robust standard errors, using clubSandwich
+  iris$cluster <- factor(rep(LETTERS[1:8], length.out = nrow(iris)))
+  standard_error_robust(
+    model,
+    vcov_type = "CR2",
+    vcov_args = list(cluster = iris$cluster)
+  )
+}
 }
