### 2020-07-24 892226b
close https://github.com/strengejacke/sjPlot/issues/652
