+++ NEWS.md
@@ -17,6 +17,7 @@
 * Allow correlation matrix to be passed to `factor_analysis()`.
 * Fix CRAN check issues.
 * Fix issue in `model_parameters()` for models with non-estimable parameters or statistics.
+* Fix issue in `model_parameters()` for *plm* models with only one parameter.
 * Fix issue in `check_heterogeneity()` in case no predictor would cause heterogeneity bias.
 * Make sure *clubSandwich* is used conditionally in all places, to properly pass CRAN checks.
 
