+++ R/model_parameters.R
@@ -26,7 +26,7 @@
 #'
 #' @note The \code{\link[=print.parameters_model]{print()}} method has several arguments to tweak the output. There is also a \href{https://easystats.github.io/see/articles/parameters.html}{\code{plot()}-method} implemented in the \href{https://easystats.github.io/see/}{\pkg{see}-package}.
 #'
-#' @details Standardization is based on \code{\link[effectsize]{standardize_parameters}}.
+#' @details Standardization is based on \code{\link[effectsize:standardize_parameters]{standardize_parameters()}}.
 #'   In case of \code{standardize = "refit"}, the data used to fit the model
 #'   will be standardized and the model is completely refitted. In such cases,
 #'   standard errors and confidence intervals refer to the standardized coefficient.
