+++ man/model_parameters.Rd
@@ -37,7 +37,7 @@ Compute and extract model parameters. See the documentation for your object's cl
  }
 }
 \details{
-Standardization is based on \code{\link[effectsize]{standardize_parameters}}.
+Standardization is based on \code{\link[effectsize:standardize_parameters]{standardize_parameters()}}.
   In case of \code{standardize = "refit"}, the data used to fit the model
   will be standardized and the model is completely refitted. In such cases,
   standard errors and confidence intervals refer to the standardized coefficient.
