+++ tests/testthat/test-model_parameters_std.R
@@ -1,6 +1,6 @@
 .runThisTest <- Sys.getenv("RunAllparametersTests") == "yes"
 
-if (.runThisTest || Sys.getenv("USER") == "travis") {
+if (.runThisTest) {
 
   if (require("testthat") &&
       require("parameters") &&
