+++ man/principal_components.Rd
@@ -76,23 +76,24 @@ There is a \code{summary()}-method that prints the Eigenvalues and (explained) v
 }
 \examples{
 library(parameters)
+if (require("psych")) {
+  principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)
+  principal_components(mtcars[, 1:7], n = 2, rotation = "oblimin",
+                       threshold = "max", sort = TRUE)
+  principal_components(mtcars[, 1:7], n = 2, threshold = 2, sort = TRUE)
 
-principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)
-principal_components(mtcars[, 1:7], n = 2, rotation = "oblimin", threshold = "max", sort = TRUE)
-principal_components(mtcars[, 1:7], n = 2, threshold = 2, sort = TRUE)
+  pca <- principal_components(mtcars[, 1:5], n = 2, rotation = "varimax")
+  summary(pca)
+  predict(pca)
 
-pca <- principal_components(mtcars[, 1:5], n = 2, rotation = "varimax")
-summary(pca)
-predict(pca)
-
-# which variables from the original data belong to which extracted component?
-closest_component(pca)
+  # which variables from the original data belong to which extracted component?
+  closest_component(pca)
 
 \donttest{
-# Automated number of components
-principal_components(mtcars[, 1:4], n = "auto")
+  # Automated number of components
+  principal_components(mtcars[, 1:4], n = "auto")
 }
-
+}
 }
 \references{
 \itemize{
