+++ man/model_parameters.merMod.Rd
@@ -101,7 +101,6 @@ if (require("lme4")) {
   model <- lmer(mpg ~ wt + (1 | gear), data = mtcars)
   model_parameters(model)
 }
-
 \donttest{
 if (require("glmmTMB")) {
   data(Salamanders)
