+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: parameters
 Title: Processing of Model Parameters
-Version: 0.8.1.1
+Version: 0.8.2
 Authors@R: 
   c(
     person("Daniel", 
