### 2020-07-26 b9421c8
Update R-check.yaml
