+++ .github/workflows/R-check.yaml
@@ -18,9 +18,9 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macOS-latest,   r: 'devel'}
+          # - {os: macOS-latest,   r: 'devel'}
           - {os: macOS-latest,   r: 'release'}
-          # - {os: macOS-latest,   r: 'oldrel'}
+          - {os: macOS-latest,   r: 'oldrel'}
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
           # - {os: windows-latest, r: 'oldrel'}
