### 2020-07-26 9cc9b14
Update R-check.yaml
