### 2020-07-26 5a2a011
fix check issues (run examples conditionally)
