### 2020-07-28 65361e9
model_parameters() for brms-meta-analysis models
