+++ NAMESPACE
@@ -383,6 +383,7 @@ S3method(plot,cluster_analysis)
 S3method(plot,equivalence_test_lm)
 S3method(plot,n_clusters)
 S3method(plot,n_factors)
+S3method(plot,parameters_brms_meta)
 S3method(plot,parameters_distribution)
 S3method(plot,parameters_efa)
 S3method(plot,parameters_model)
@@ -405,6 +406,7 @@ S3method(print,cluster_discrimintation)
 S3method(print,equivalence_test_lm)
 S3method(print,n_clusters)
 S3method(print,n_factors)
+S3method(print,parameters_brms_meta)
 S3method(print,parameters_clusters)
 S3method(print,parameters_distribution)
 S3method(print,parameters_efa)
@@ -740,6 +742,7 @@ importFrom(insight,get_data)
 importFrom(insight,get_parameters)
 importFrom(insight,get_priors)
 importFrom(insight,get_random)
+importFrom(insight,get_response)
 importFrom(insight,get_statistic)
 importFrom(insight,get_varcov)
 importFrom(insight,get_variance)
