### 2020-07-30 f9bc6f9
Update R-check.yaml
