+++ .github/workflows/R-check.yaml
@@ -18,12 +18,12 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macOS-latest,   r: 'devel'}
+          # - {os: macOS-latest,   r: 'devel'}
           - {os: macOS-latest,   r: 'release'}
-          # - {os: macOS-latest,   r: 'oldrel'}
+          - {os: macOS-latest,   r: 'oldrel'}
           - {os: windows-latest, r: 'devel'}
           - {os: windows-latest, r: 'release'}
-          # - {os: windows-latest, r: 'oldrel'}
+          - {os: windows-latest, r: 'oldrel'}
           - {os: ubuntu-16.04,   r: 'devel', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
           - {os: ubuntu-16.04,   r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
           # - {os: ubuntu-16.04,   r: 'oldrel',  rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
