### 2020-07-30 3f148aa
Preliminary support for `semLME` (*smicd*)
