+++ R/utils.R
@@ -253,3 +253,10 @@
   }
   model
 }
+
+
+
+
+.is_semLme <- function(x) {
+  all(inherits(x, c("sem", "lme")))
+}
