+++ R/standard_error.R
@@ -856,6 +856,22 @@ standard_error.glht <- function(model, ...) {
 }
 
 
+#' @export
+standard_error.sem <- function(model, ...) {
+  if (!.is_semLme(model)) {
+    return(NULL)
+  }
+  if (is.null(model$se)) {
+    warning("Model has no standard errors. Please fit model again with bootstrapped standard errors.", call. = FALSE)
+    return(NULL)
+  }
+  .data_frame(
+    Parameter = names(model$se),
+    SE = unname(model$se)
+  )
+}
+
+
 #' @importFrom stats na.omit
 #' @export
 standard_error.robmixglm <- function(model, ...) {
