+++ NAMESPACE
@@ -94,6 +94,7 @@ S3method(ci,rma)
 S3method(ci,rms)
 S3method(ci,rq)
 S3method(ci,rqss)
+S3method(ci,sem)
 S3method(ci,speedglm)
 S3method(ci,speedlm)
 S3method(ci,survreg)
@@ -225,6 +226,7 @@ S3method(model_parameters,probitmfx)
 S3method(model_parameters,rlmerMod)
 S3method(model_parameters,rma)
 S3method(model_parameters,rqss)
+S3method(model_parameters,sem)
 S3method(model_parameters,stanfit)
 S3method(model_parameters,stanmvreg)
 S3method(model_parameters,stanreg)
@@ -355,6 +357,7 @@ S3method(p_value,rms)
 S3method(p_value,robmixglm)
 S3method(p_value,rq)
 S3method(p_value,rqss)
+S3method(p_value,sem)
 S3method(p_value,speedlm)
 S3method(p_value,stanreg)
 S3method(p_value,survreg)
@@ -608,6 +611,7 @@ S3method(standard_error,rms)
 S3method(standard_error,robmixglm)
 S3method(standard_error,rq)
 S3method(standard_error,rqss)
+S3method(standard_error,sem)
 S3method(standard_error,stanreg)
 S3method(standard_error,survreg)
 S3method(standard_error,svyglm)
