+++ NAMESPACE
@@ -589,6 +589,8 @@ S3method(standard_error,lrm)
 S3method(standard_error,merMod)
 S3method(standard_error,metaplus)
 S3method(standard_error,mixor)
+S3method(standard_error,mle)
+S3method(standard_error,mle2)
 S3method(standard_error,mlm)
 S3method(standard_error,multinom)
 S3method(standard_error,mvstanreg)
