+++ NAMESPACE
@@ -77,6 +77,8 @@ S3method(ci,lrm)
 S3method(ci,merMod)
 S3method(ci,metaplus)
 S3method(ci,mixor)
+S3method(ci,mle)
+S3method(ci,mle2)
 S3method(ci,mlm)
 S3method(ci,multinom)
 S3method(ci,negbin)
@@ -205,14 +207,18 @@ S3method(model_parameters,kmeans)
 S3method(model_parameters,lavaan)
 S3method(model_parameters,list)
 S3method(model_parameters,lme)
+S3method(model_parameters,logistf)
 S3method(model_parameters,logitmfx)
 S3method(model_parameters,logitor)
 S3method(model_parameters,mcmc)
 S3method(model_parameters,merMod)
 S3method(model_parameters,metaplus)
 S3method(model_parameters,mixor)
+S3method(model_parameters,mle)
+S3method(model_parameters,mle2)
 S3method(model_parameters,mlm)
 S3method(model_parameters,multinom)
+S3method(model_parameters,negbin)
 S3method(model_parameters,negbinirr)
 S3method(model_parameters,negbinmfx)
 S3method(model_parameters,omega)
@@ -334,6 +340,7 @@ S3method(p_value,maxLik)
 S3method(p_value,merMod)
 S3method(p_value,metaplus)
 S3method(p_value,mixor)
+S3method(p_value,mle2)
 S3method(p_value,mlm)
 S3method(p_value,multinom)
 S3method(p_value,negbin)
