+++ R/standard_error.R
@@ -847,7 +847,7 @@ standard_error.bayesx <- function(model, ...) {
 
 
 #' @export
-standard_error.mle2 <- function(x, ...) {
+standard_error.mle2 <- function(model, ...) {
   if (!requireNamespace("bbmle", quietly = TRUE)) {
     stop("Package `bbmle` needs to be installed to extract standard errors.", call. = FALSE)
   }
