+++ R/p_value.R
@@ -756,7 +756,7 @@ p_value.betamfx <- function(model, component = c("all", "conditional", "precisio
 
 
 #' @export
-p_value.mle2 <- function(x, ...) {
+p_value.mle2 <- function(model, ...) {
   if (!requireNamespace("bbmle", quietly = TRUE)) {
     stop("Package `bbmle` needs to be installed to extract p-values.", call. = FALSE)
   }
