### 2020-08-03 46387c7
add model_parameters for "mira"
