+++ R/model_parameters.R
@@ -16,6 +16,7 @@
 #'  \item{\link[=model_parameters.kmeans]{Cluster models} (k-means, ...)}
 #'  \item{\link[=model_parameters.rma]{Meta-Analysis via linear (mixed) models} (\code{rma})}
 #'  \item{\link[=model_parameters.glht]{Hypothesis Testing} (\code{glht})}
+#'  \item{\link[=model_parameters.mira]{Multiply imputed repeated analyses} (\code{mira})}
 #'  }
 #'
 #' @param model Statistical Model.
