+++ man/model_parameters.mira.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/model_parameters.mira.R
+\name{model_parameters.mira}
+\alias{model_parameters.mira}
+\title{Parameters from multiply imputed repeated analyses}
+\usage{
+\method{model_parameters}{mira}(model, ci = 0.95, exponentiate = FALSE, p_adjust = NULL, ...)
+}
+\arguments{
+\item{model}{An object of class \code{mira}.}
+
+\item{...}{Arguments passed to or from other methods.}
+}
+\description{
+Format models of class \code{mira}, obtained from \code{mice::width.mids()}.
+}
+\examples{
+library(parameters)
+if (require("mice")) {
+  data(nhanes2)
+  imp <- mice(nhanes2)
+  fit <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
+  model_parameters(fit)
+}
+}
