+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New supported model classes
 
 * Preliminary support for `semLME` (*smicd*), `mle2` (*bbmle*), `mle` (*stats4*)
+* `model_parameters()` for objects of class `mira` (*mice*).
 
 ## Changes to functions
 
