+++ NAMESPACE
@@ -155,6 +155,7 @@ S3method(equivalence_test,zeroinfl)
 S3method(factor_analysis,data.frame)
 S3method(format_parameters,default)
 S3method(format_parameters,glmm)
+S3method(format_parameters,mira)
 S3method(format_parameters,parameters_model)
 S3method(format_parameters,rma)
 S3method(kurtosis,data.frame)
@@ -213,6 +214,7 @@ S3method(model_parameters,logitor)
 S3method(model_parameters,mcmc)
 S3method(model_parameters,merMod)
 S3method(model_parameters,metaplus)
+S3method(model_parameters,mira)
 S3method(model_parameters,mixor)
 S3method(model_parameters,mle)
 S3method(model_parameters,mle2)
