+++ R/standard_error.R
@@ -851,7 +851,7 @@ standard_error.mle2 <- function(model, ...) {
   if (!requireNamespace("bbmle", quietly = TRUE)) {
     stop("Package `bbmle` needs to be installed to extract standard errors.", call. = FALSE)
   }
-  s <- bbmle::summary(x)
+  s <- bbmle::summary(model)
   .data_frame(
     Parameter = names(s@coef[, 2]),
     SE = unname(s@coef[, 2])
