+++ _pkgdown.yml
@@ -144,6 +144,8 @@ navbar:
         href: articles/model_parameters_standardized.html
       - text: Robust Estimation of Standard Errors
         href: articles/model_parameters_robust.html
+      - text: Model Parameters and Missing Data
+        href: articles/model_parameters_mice.html
       - text: Panel Data Analysis
         href: articles/demean.html
   - text: Parameters Selection
