+++ NAMESPACE
@@ -76,6 +76,7 @@ S3method(ci,logitor)
 S3method(ci,lrm)
 S3method(ci,merMod)
 S3method(ci,metaplus)
+S3method(ci,mipo)
 S3method(ci,mixor)
 S3method(ci,mle)
 S3method(ci,mle2)
@@ -124,6 +125,7 @@ S3method(degrees_of_freedom,emmGrid)
 S3method(degrees_of_freedom,glht)
 S3method(degrees_of_freedom,logitmfx)
 S3method(degrees_of_freedom,logitor)
+S3method(degrees_of_freedom,mipo)
 S3method(degrees_of_freedom,negbinirr)
 S3method(degrees_of_freedom,negbinmfx)
 S3method(degrees_of_freedom,poissonirr)
@@ -214,6 +216,7 @@ S3method(model_parameters,logitor)
 S3method(model_parameters,mcmc)
 S3method(model_parameters,merMod)
 S3method(model_parameters,metaplus)
+S3method(model_parameters,mipo)
 S3method(model_parameters,mira)
 S3method(model_parameters,mixor)
 S3method(model_parameters,mle)
@@ -341,6 +344,7 @@ S3method(p_value,lrm)
 S3method(p_value,maxLik)
 S3method(p_value,merMod)
 S3method(p_value,metaplus)
+S3method(p_value,mipo)
 S3method(p_value,mixor)
 S3method(p_value,mle2)
 S3method(p_value,mlm)
@@ -597,6 +601,7 @@ S3method(standard_error,logitor)
 S3method(standard_error,lrm)
 S3method(standard_error,merMod)
 S3method(standard_error,metaplus)
+S3method(standard_error,mipo)
 S3method(standard_error,mixor)
 S3method(standard_error,mle)
 S3method(standard_error,mle2)
