### 2020-08-03 8e8307a
docs / vignette
