+++ man/model_parameters.mira.Rd
@@ -47,4 +47,14 @@ if (require("mice") && require("gee")) {
 
   model_parameters(fit)
 }
+
+# and it works with pooled results
+if (require("mice")) {
+  data("nhanes2")
+  imp <- mice(nhanes2)
+  fit <- with(data = imp, exp = lm(bmi ~ age + hyp + chl))
+  pooled <- pool(fit)
+
+  model_parameters(pooled)
+}
 }
