+++ R/check_heterogeneity.R
@@ -4,7 +4,7 @@
 #' @export
 check_heterogeneity <- function(x, select = NULL, group = NULL) {
   if (insight::is_model(x)) {
-    group <- insight::find_random(x, split_nested = TRUE, flatten = FALSE)
+    group <- insight::find_random(x, split_nested = TRUE, flatten = TRUE)
     if (is.null(group)) {
       stop("Model is no mixed model. Please provide a mixed model, or a data frame and arguments 'select' and 'group'.")
     }
@@ -32,7 +32,7 @@ check_heterogeneity <- function(x, select = NULL, group = NULL) {
 
     # check if any within-variable differs from zero. if yes, we have
     # a within-subject effect
-    if (any(sum(abs(d[[within_name]]) > 1e-5) > 0)) {
+    if (any(sum(abs(d[[within_name]]) > 1e-5, na.rm = TRUE) > 0)) {
       predictor
     } else {
       NULL
