+++ NEWS.md
@@ -16,6 +16,8 @@
 ## Bug fixes
 
 * Fix issue in `model_parameters.metaplus()` with random effects.
+* Fix issue in `check_heterogeneity()` when `x` was a mixed model.
+* Fix issue in `check_heterogeneity()` for data with missing values.
 
 # parameters 0.8.2
 
