### 2020-08-13 264f3f5
model_parameters for emmGrid-contrast
