### 2020-08-13 85b91e8
allow shortcuts for emmGrid
