### 2020-08-14 #2445 @Ohlsen
Hi, @sbenthall, apparently i missed some notification and am also not able to edit the checklist. Could you start the invitation again? Sorry for the trouble my self-assignment caused. I´m new to this as well ...
