### 2020-08-15 #2445 @Ohlsen
I´m still not able to edit the checklist. Perhaps it might work if i´m re-invited and not only re-assigned?
