### 2020-08-16 fd3e283
#283 (make sem more robust)
