+++ R/standardize_names.R
@@ -58,7 +58,7 @@ standardize_names.parameters_model <- function(data, style = c("easystats", "bro
   cn <- colnames(data)
 
   if (style == "easystats") {
-    cn[cn %in% c("t", "z", "F", "chisq", "t / F", "z / Chisq")] <- "Statistic"
+    cn[cn %in% c("t", "z", "F", "chisq", "chi-sq", "t / F", "z / Chisq")] <- "Statistic"
     cn[cn %in% c("Median", "Mean", "MAP")] <- "Coefficient"
     cn[cn %in% c("df_residual", "df_error")] <- "df"
   } else {
@@ -74,7 +74,7 @@ standardize_names.parameters_model <- function(data, style = c("easystats", "bro
     # more sophisticated replacements
     cn[cn %in% c("df_residual", "df_error")] <- "df.error"
     cn[cn %in% c("Coefficient", "Std_Coefficient", "Median", "Mean", "MAP")] <- "estimate"
-    cn[cn %in% c("t", "z", "F", "chisq", "t / F", "z / Chisq")] <- "statistic"
+    cn[cn %in% c("t", "z", "F", "chisq", "chi-sq", "t / F", "z / Chisq")] <- "statistic"
     # fancy regex replacements
     cn <- gsub("^CI_low", "conf.low", cn)
     cn <- gsub("^CI_high", "conf.high", cn)
