### 2020-08-18 2a40d5e
update news
