+++ NAMESPACE
@@ -73,6 +73,8 @@ S3method(ci,lme)
 S3method(ci,logistf)
 S3method(ci,logitmfx)
 S3method(ci,logitor)
+S3method(ci,lqm)
+S3method(ci,lqmm)
 S3method(ci,lrm)
 S3method(ci,merMod)
 S3method(ci,metaplus)
@@ -213,6 +215,8 @@ S3method(model_parameters,lme)
 S3method(model_parameters,logistf)
 S3method(model_parameters,logitmfx)
 S3method(model_parameters,logitor)
+S3method(model_parameters,lqm)
+S3method(model_parameters,lqmm)
 S3method(model_parameters,mcmc)
 S3method(model_parameters,merMod)
 S3method(model_parameters,metaplus)
@@ -340,6 +344,8 @@ S3method(p_value,lmerMod)
 S3method(p_value,logistf)
 S3method(p_value,logitmfx)
 S3method(p_value,logitor)
+S3method(p_value,lqm)
+S3method(p_value,lqmm)
 S3method(p_value,lrm)
 S3method(p_value,maxLik)
 S3method(p_value,merMod)
@@ -598,6 +604,8 @@ S3method(standard_error,lme)
 S3method(standard_error,logistf)
 S3method(standard_error,logitmfx)
 S3method(standard_error,logitor)
+S3method(standard_error,lqm)
+S3method(standard_error,lqmm)
 S3method(standard_error,lrm)
 S3method(standard_error,merMod)
 S3method(standard_error,metaplus)
