### 2020-08-19 8aa1f8b
also include df in model_parameters.lqmm
