+++ NAMESPACE
@@ -127,6 +127,8 @@ S3method(degrees_of_freedom,emmGrid)
 S3method(degrees_of_freedom,glht)
 S3method(degrees_of_freedom,logitmfx)
 S3method(degrees_of_freedom,logitor)
+S3method(degrees_of_freedom,lqm)
+S3method(degrees_of_freedom,lqmm)
 S3method(degrees_of_freedom,mipo)
 S3method(degrees_of_freedom,negbinirr)
 S3method(degrees_of_freedom,negbinmfx)
