### 2020-08-26 easystats/parameters#290 @Ohlsen
Include self defined parameters like indirect effects for lavaan::sem() model output
First of all, this is more of a feature request, not a exaclty a blocker for acceptance.

In the JOSS Paper lavaan::sem() is listed in the image of modeling methods supported  by the parameters-package. The vignette https://stat.ethz.ch/CRAN/web/packages/parameters/vignettes/efa_cfa.html
offers an explanation how parameters() works for EFA and CFA, but keeps quiet of lavaan::sem(). Structural equation models have an enormous amount of modeling possibilities. For example, it is possible to estimate indirect effects in partial mediation structures. For this, the effects have to be defined by the ':=' operator. At the moment, parameters does not extract those.

```r
library(lavaan)
library(parameters)

set.seed(1234)
X <- rnorm(100)
M <- 0.5*X + rnorm(100)
Y <- 0.7*M + rnorm(100)
Data <- data.frame(X = X, Y = Y, M = M)
model <- ' # direct effect
             Y ~ c*X
           # mediator
             M ~ a*X
             Y ~ b*M
           # indirect effect (a*b)
             ab := a*b
           # total effect
             total := c + (a*b)
         '
fit <- sem(model, data = Data, test = "Satorra-Bentler")

summary(fit)
model_parameters(fit, standardize = TRUE)
```

I think, having the possibility to include those effects in the parameters-package would be a useful addition. Concerning the image in the JOSS-paper with the supported models, i would suggest to empasize the support of lavaan::cfa() instead of lavaan::sem() at the moment, because the vignette is pretty silent about lavaan::sem() and some output ist not yet available in the parameters package.
