### 2020-08-26 easystats/parameters#290 @strengejacke
I added the "definition" parameters now as well:

``` r
library(lavaan)
#> This is lavaan 0.6-7
#> lavaan is BETA software! Please report any bugs.
library(parameters)

set.seed(1234)
X <- rnorm(100)
M <- 0.5*X + rnorm(100)
Y <- 0.7*M + rnorm(100)
Data <- data.frame(X = X, Y = Y, M = M)
model <- ' # direct effect
             Y ~ c*X
           # mediator
             M ~ a*X
             Y ~ b*M
           # indirect effect (a*b)
             ab := a*b
           # total effect
             total := c + (a*b)
         '
fit <- sem(model, data = Data, test = "Satorra-Bentler")

summary(fit)
#> lavaan 0.6-7 ended normally after 12 iterations
#>
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of free parameters                          5
#>
#>   Number of observations                           100
#>
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                                 0.000       0.000
#>   Degrees of freedom                                 0           0
#>
#> Parameter Estimates:
#>
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#>
#> Regressions:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   Y ~
#>     X          (c)    0.036    0.104    0.348    0.728
#>   M ~
#>     X          (a)    0.474    0.103    4.613    0.000
#>   Y ~
#>     M          (b)    0.788    0.092    8.539    0.000
#>
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .Y                 0.898    0.127    7.071    0.000
#>    .M                 1.054    0.149    7.071    0.000
#>
#> Defined Parameters:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>     ab                0.374    0.092    4.059    0.000
#>     total             0.410    0.125    3.287    0.001

model_parameters(fit, standardize = TRUE)
#> # Regression type
#>
#> Link  | Coefficient |   SE |        95% CI |      p
#> ---------------------------------------------------
#> Y ~ X |        0.03 | 0.08 | [-0.13, 0.18] | 0.728
#> M ~ X |        0.42 | 0.08 | [ 0.26, 0.57] | < .001
#> Y ~ M |        0.68 | 0.06 | [ 0.56, 0.80] | < .001
#>
#> # Defined type
#>
#> Link             | Coefficient |   SE |       95% CI |      p
#> -------------------------------------------------------------
#> ab := a*b        |        0.28 | 0.06 | [0.16, 0.41] | < .001
#> total := c+(a*b) |        0.31 | 0.09 | [0.14, 0.48] | < .001
```

Not sure, though, if the output still lacks some useful information? Is there more you would add to the output?
