### 2020-08-26 28958de
improve print for SEM (#290)
