### 2020-08-26 910e588
panelr is removed from CRAN (again)
