### 2020-08-26 e0f855f
typos, add example to vignette
