+++ vignettes/efa_cfa.Rmd
@@ -176,4 +176,30 @@ performance::compare_performance(big5, big6)
 All in all, it seems that the big 5 structure remains quite reliable.
 
 
+
+# Structural Equation Modeling
+
+The last example shows one of the enormous amount of modeling possibilities for structural equation models, in particular an example for mediation analysis, i.e. a model that estimates indirect effects in partial mediation structures.
+
+```{r message=FALSE, warning=FALSE}
+set.seed(1234)
+X <- rnorm(100)
+M <- 0.5 * X + rnorm(100)
+Y <- 0.7 * M + rnorm(100)
+Data <- data.frame(X = X, Y = Y, M = M)
+
+model <- ' # direct effect
+             Y ~ c*X
+           # mediator
+             M ~ a*X
+             Y ~ b*M
+           # indirect effect (a*b)
+             ab := a*b
+           # total effect
+             total := c + (a*b)
+         '
+fit <- sem(model, data = Data, test = "Satorra-Bentler")
+model_parameters(fit)
+```
+
 # References
