### 2020-08-26 easystats/parameters#291 @Ohlsen
Paper: Consider adding std.lv parameter estimations to unstandardized and completely standardized solution
In lavaan::sem() each estimated parameter can be inspected from an unstandardized solution, a completely standardized solution (std.all) or a solution, where only the latent variables are standardized (std.lv).
In my own research i never needed std.lv, but i guess to offer a complete output you might consider adding this.

`
```
model <- "
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
"

fit <- sem(model, data=PoliticalDemocracy, test = "Satorra-Bentler")

# compare lavaan output to model_parameters()
summary(fit, standardized=TRUE)

model_parameters(fit)
model_parameters(fit, standardize = TRUE)
```
