### 2020-08-26 easystats/parameters#291 @strengejacke
Thank you for this suggestion. We have added all standardize-options to `model_parameters()` and also added tests for the new features. Here's how `model_parameters()` works now:

``` r
library(lavaan)
#> This is lavaan 0.6-7
#> lavaan is BETA software! Please report any bugs.
library(parameters)

model <- "
  # measurement model
    ind60 =~ x1 + x2 + x3
    dem60 =~ y1 + y2 + y3 + y4
    dem65 =~ y5 + y6 + y7 + y8
  # regressions
    dem60 ~ ind60
    dem65 ~ ind60 + dem60
  # residual correlations
    y1 ~~ y5
    y2 ~~ y4 + y6
    y3 ~~ y7
    y4 ~~ y8
    y6 ~~ y8
"

fit <- sem(model, data=PoliticalDemocracy, test = "Satorra-Bentler")

lavaan::standardizedsolution(fit, se = TRUE, level = .95, type = "std.all")
#>      lhs op   rhs est.std    se      z pvalue ci.lower ci.upper
#> 1  ind60 =~    x1   0.920 0.023 40.005  0.000    0.875    0.965
#> 2  ind60 =~    x2   0.973 0.016 59.110  0.000    0.941    1.005
#> 3  ind60 =~    x3   0.872 0.031 28.081  0.000    0.811    0.933
#> 4  dem60 =~    y1   0.850 0.042 20.200  0.000    0.768    0.933
#> 5  dem60 =~    y2   0.717 0.064 11.157  0.000    0.591    0.843
#> 6  dem60 =~    y3   0.722 0.063 11.439  0.000    0.599    0.846
#> 7  dem60 =~    y4   0.846 0.043 19.462  0.000    0.761    0.931
#> 8  dem65 =~    y5   0.808 0.048 16.847  0.000    0.714    0.902
#> 9  dem65 =~    y6   0.746 0.058 12.924  0.000    0.633    0.859
#> 10 dem65 =~    y7   0.824 0.045 18.392  0.000    0.736    0.911
#> 11 dem65 =~    y8   0.828 0.045 18.477  0.000    0.740    0.916
#> 12 dem60  ~ ind60   0.447 0.103  4.323  0.000    0.244    0.649
#> 13 dem65  ~ ind60   0.182 0.070  2.587  0.010    0.044    0.320
#> 14 dem65  ~ dem60   0.885 0.051 17.405  0.000    0.786    0.985
#> 15    y1 ~~    y5   0.296 0.140  2.117  0.034    0.022    0.570
#> 16    y2 ~~    y4   0.273 0.120  2.265  0.024    0.037    0.508
#> 17    y2 ~~    y6   0.356 0.097  3.690  0.000    0.167    0.545
#> 18    y3 ~~    y7   0.191 0.133  1.432  0.152   -0.070    0.452
#> 19    y4 ~~    y8   0.109 0.131  0.832  0.405   -0.147    0.365
#> 20    y6 ~~    y8   0.338 0.110  3.079  0.002    0.123    0.553
#> 21    x1 ~~    x1   0.154 0.042  3.637  0.000    0.071    0.237
#> 22    x2 ~~    x2   0.053 0.032  1.661  0.097   -0.010    0.116
#> 23    x3 ~~    x3   0.239 0.054  4.419  0.000    0.133    0.346
#> 24    y1 ~~    y1   0.277 0.072  3.865  0.000    0.136    0.417
#> 25    y2 ~~    y2   0.486 0.092  5.269  0.000    0.305    0.666
#> 26    y3 ~~    y3   0.478 0.091  5.242  0.000    0.299    0.657
#> 27    y4 ~~    y4   0.285 0.073  3.875  0.000    0.141    0.429
#> 28    y5 ~~    y5   0.347 0.078  4.478  0.000    0.195    0.499
#> 29    y6 ~~    y6   0.443 0.086  5.149  0.000    0.275    0.612
#> 30    y7 ~~    y7   0.322 0.074  4.359  0.000    0.177    0.466
#> 31    y8 ~~    y8   0.315 0.074  4.242  0.000    0.169    0.460
#> 32 ind60 ~~ ind60   1.000 0.000     NA     NA    1.000    1.000
#> 33 dem60 ~~ dem60   0.800 0.092  8.669  0.000    0.619    0.981
#> 34 dem65 ~~ dem65   0.039 0.048  0.808  0.419   -0.056    0.134
model_parameters(fit, standardize = TRUE)
#> # Loading
#>
#> Link        | Coefficient |   SE |       95% CI |      p
#> --------------------------------------------------------
#> ind60 =~ x1 |        0.92 | 0.02 | [0.87, 0.96] | < .001
#> ind60 =~ x2 |        0.97 | 0.02 | [0.94, 1.01] | < .001
#> ind60 =~ x3 |        0.87 | 0.03 | [0.81, 0.93] | < .001
#> dem60 =~ y1 |        0.85 | 0.04 | [0.77, 0.93] | < .001
#> dem60 =~ y2 |        0.72 | 0.06 | [0.59, 0.84] | < .001
#> dem60 =~ y3 |        0.72 | 0.06 | [0.60, 0.85] | < .001
#> dem60 =~ y4 |        0.85 | 0.04 | [0.76, 0.93] | < .001
#> dem65 =~ y5 |        0.81 | 0.05 | [0.71, 0.90] | < .001
#> dem65 =~ y6 |        0.75 | 0.06 | [0.63, 0.86] | < .001
#> dem65 =~ y7 |        0.82 | 0.04 | [0.74, 0.91] | < .001
#> dem65 =~ y8 |        0.83 | 0.04 | [0.74, 0.92] | < .001
#>
#> # Regression
#>
#> Link          | Coefficient |   SE |       95% CI |      p
#> ----------------------------------------------------------
#> dem60 ~ ind60 |        0.45 | 0.10 | [0.24, 0.65] | < .001
#> dem65 ~ ind60 |        0.18 | 0.07 | [0.04, 0.32] | 0.010
#> dem65 ~ dem60 |        0.89 | 0.05 | [0.79, 0.98] | < .001
#>
#> # Correlation
#>
#> Link     | Coefficient |   SE |        95% CI |      p
#> ------------------------------------------------------
#> y1 ~~ y5 |        0.30 | 0.14 | [ 0.02, 0.57] | 0.034
#> y2 ~~ y4 |        0.27 | 0.12 | [ 0.04, 0.51] | 0.024
#> y2 ~~ y6 |        0.36 | 0.10 | [ 0.17, 0.55] | < .001
#> y3 ~~ y7 |        0.19 | 0.13 | [-0.07, 0.45] | 0.152
#> y4 ~~ y8 |        0.11 | 0.13 | [-0.15, 0.37] | 0.405
#> y6 ~~ y8 |        0.34 | 0.11 | [ 0.12, 0.55] | 0.002

lavaan::standardizedsolution(fit, se = TRUE, level = .95, type = "std.lv")
#>      lhs op   rhs est.std    se      z pvalue ci.lower ci.upper
#> 1  ind60 =~    x1   0.670 0.065 10.346  0.000    0.543    0.797
#> 2  ind60 =~    x2   1.460 0.128 11.424  0.000    1.210    1.711
#> 3  ind60 =~    x3   1.218 0.128  9.480  0.000    0.966    1.470
#> 4  dem60 =~    y1   2.223 0.251  8.844  0.000    1.730    2.716
#> 5  dem60 =~    y2   2.794 0.407  6.864  0.000    1.996    3.592
#> 6  dem60 =~    y3   2.351 0.338  6.955  0.000    1.689    3.014
#> 7  dem60 =~    y4   2.812 0.322  8.735  0.000    2.181    3.443
#> 8  dem65 =~    y5   2.103 0.256  8.217  0.000    1.601    2.604
#> 9  dem65 =~    y6   2.493 0.340  7.328  0.000    1.826    3.160
#> 10 dem65 =~    y7   2.691 0.317  8.481  0.000    2.069    3.312
#> 11 dem65 =~    y8   2.662 0.313  8.512  0.000    2.049    3.275
#> 12 dem60  ~ ind60   0.447 0.103  4.323  0.000    0.244    0.649
#> 13 dem65  ~ ind60   0.182 0.070  2.587  0.010    0.044    0.320
#> 14 dem65  ~ dem60   0.885 0.051 17.405  0.000    0.786    0.985
#> 15    y1 ~~    y5   0.624 0.358  1.741  0.082   -0.079    1.326
#> 16    y2 ~~    y4   1.313 0.702  1.871  0.061   -0.063    2.689
#> 17    y2 ~~    y6   2.153 0.734  2.934  0.003    0.715    3.591
#> 18    y3 ~~    y7   0.795 0.608  1.308  0.191   -0.396    1.986
#> 19    y4 ~~    y8   0.348 0.442  0.787  0.431   -0.519    1.215
#> 20    y6 ~~    y8   1.356 0.568  2.386  0.017    0.242    2.470
#> 21    x1 ~~    x1   0.082 0.019  4.184  0.000    0.043    0.120
#> 22    x2 ~~    x2   0.120 0.070  1.718  0.086   -0.017    0.256
#> 23    x3 ~~    x3   0.467 0.090  5.177  0.000    0.290    0.643
#> 24    y1 ~~    y1   1.891 0.444  4.256  0.000    1.020    2.762
#> 25    y2 ~~    y2   7.373 1.374  5.366  0.000    4.680   10.066
#> 26    y3 ~~    y3   5.067 0.952  5.324  0.000    3.202    6.933
#> 27    y4 ~~    y4   3.148 0.739  4.261  0.000    1.700    4.596
#> 28    y5 ~~    y5   2.351 0.480  4.895  0.000    1.410    3.292
#> 29    y6 ~~    y6   4.954 0.914  5.419  0.000    3.162    6.746
#> 30    y7 ~~    y7   3.431 0.713  4.814  0.000    2.034    4.829
#> 31    y8 ~~    y8   3.254 0.695  4.685  0.000    1.893    4.615
#> 32 ind60 ~~ ind60   1.000 0.000     NA     NA    1.000    1.000
#> 33 dem60 ~~ dem60   0.800 0.092  8.669  0.000    0.619    0.981
#> 34 dem65 ~~ dem65   0.039 0.048  0.808  0.419   -0.056    0.134
model_parameters(fit, standardize = "latent")
#> # Loading
#>
#> Link        | Coefficient |   SE |       95% CI |      p
#> --------------------------------------------------------
#> ind60 =~ x1 |        0.67 | 0.06 | [0.54, 0.80] | < .001
#> ind60 =~ x2 |        1.46 | 0.13 | [1.21, 1.71] | < .001
#> ind60 =~ x3 |        1.22 | 0.13 | [0.97, 1.47] | < .001
#> dem60 =~ y1 |        2.22 | 0.25 | [1.73, 2.72] | < .001
#> dem60 =~ y2 |        2.79 | 0.41 | [2.00, 3.59] | < .001
#> dem60 =~ y3 |        2.35 | 0.34 | [1.69, 3.01] | < .001
#> dem60 =~ y4 |        2.81 | 0.32 | [2.18, 3.44] | < .001
#> dem65 =~ y5 |        2.10 | 0.26 | [1.60, 2.60] | < .001
#> dem65 =~ y6 |        2.49 | 0.34 | [1.83, 3.16] | < .001
#> dem65 =~ y7 |        2.69 | 0.32 | [2.07, 3.31] | < .001
#> dem65 =~ y8 |        2.66 | 0.31 | [2.05, 3.28] | < .001
#>
#> # Regression
#>
#> Link          | Coefficient |   SE |       95% CI |      p
#> ----------------------------------------------------------
#> dem60 ~ ind60 |        0.45 | 0.10 | [0.24, 0.65] | < .001
#> dem65 ~ ind60 |        0.18 | 0.07 | [0.04, 0.32] | 0.010
#> dem65 ~ dem60 |        0.89 | 0.05 | [0.79, 0.98] | < .001
#>
#> # Correlation
#>
#> Link     | Coefficient |   SE |        95% CI |     p
#> -----------------------------------------------------
#> y1 ~~ y5 |        0.62 | 0.36 | [-0.08, 1.33] | 0.082
#> y2 ~~ y4 |        1.31 | 0.70 | [-0.06, 2.69] | 0.061
#> y2 ~~ y6 |        2.15 | 0.73 | [ 0.71, 3.59] | 0.003
#> y3 ~~ y7 |        0.79 | 0.61 | [-0.40, 1.99] | 0.191
#> y4 ~~ y8 |        0.35 | 0.44 | [-0.52, 1.22] | 0.431
#> y6 ~~ y8 |        1.36 | 0.57 | [ 0.24, 2.47] | 0.017

lavaan::standardizedsolution(fit, se = TRUE, level = .95, type = "std.nox")
#>      lhs op   rhs est.std    se      z pvalue ci.lower ci.upper
#> 1  ind60 =~    x1   0.920 0.023 40.005  0.000    0.875    0.965
#> 2  ind60 =~    x2   0.973 0.016 59.110  0.000    0.941    1.005
#> 3  ind60 =~    x3   0.872 0.031 28.081  0.000    0.811    0.933
#> 4  dem60 =~    y1   0.850 0.042 20.200  0.000    0.768    0.933
#> 5  dem60 =~    y2   0.717 0.064 11.157  0.000    0.591    0.843
#> 6  dem60 =~    y3   0.722 0.063 11.439  0.000    0.599    0.846
#> 7  dem60 =~    y4   0.846 0.043 19.462  0.000    0.761    0.931
#> 8  dem65 =~    y5   0.808 0.048 16.847  0.000    0.714    0.902
#> 9  dem65 =~    y6   0.746 0.058 12.924  0.000    0.633    0.859
#> 10 dem65 =~    y7   0.824 0.045 18.392  0.000    0.736    0.911
#> 11 dem65 =~    y8   0.828 0.045 18.477  0.000    0.740    0.916
#> 12 dem60  ~ ind60   0.447 0.103  4.323  0.000    0.244    0.649
#> 13 dem65  ~ ind60   0.182 0.070  2.587  0.010    0.044    0.320
#> 14 dem65  ~ dem60   0.885 0.051 17.405  0.000    0.786    0.985
#> 15    y1 ~~    y5   0.296 0.140  2.117  0.034    0.022    0.570
#> 16    y2 ~~    y4   0.273 0.120  2.265  0.024    0.037    0.508
#> 17    y2 ~~    y6   0.356 0.097  3.690  0.000    0.167    0.545
#> 18    y3 ~~    y7   0.191 0.133  1.432  0.152   -0.070    0.452
#> 19    y4 ~~    y8   0.109 0.131  0.832  0.405   -0.147    0.365
#> 20    y6 ~~    y8   0.338 0.110  3.079  0.002    0.123    0.553
#> 21    x1 ~~    x1   0.154 0.042  3.637  0.000    0.071    0.237
#> 22    x2 ~~    x2   0.053 0.032  1.661  0.097   -0.010    0.116
#> 23    x3 ~~    x3   0.239 0.054  4.419  0.000    0.133    0.346
#> 24    y1 ~~    y1   0.277 0.072  3.865  0.000    0.136    0.417
#> 25    y2 ~~    y2   0.486 0.092  5.269  0.000    0.305    0.666
#> 26    y3 ~~    y3   0.478 0.091  5.242  0.000    0.299    0.657
#> 27    y4 ~~    y4   0.285 0.073  3.875  0.000    0.141    0.429
#> 28    y5 ~~    y5   0.347 0.078  4.478  0.000    0.195    0.499
#> 29    y6 ~~    y6   0.443 0.086  5.149  0.000    0.275    0.612
#> 30    y7 ~~    y7   0.322 0.074  4.359  0.000    0.177    0.466
#> 31    y8 ~~    y8   0.315 0.074  4.242  0.000    0.169    0.460
#> 32 ind60 ~~ ind60   1.000 0.000     NA     NA    1.000    1.000
#> 33 dem60 ~~ dem60   0.800 0.092  8.669  0.000    0.619    0.981
#> 34 dem65 ~~ dem65   0.039 0.048  0.808  0.419   -0.056    0.134
model_parameters(fit, standardize = "no_exogenous")
#> # Loading
#>
#> Link        | Coefficient |   SE |       95% CI |      p
#> --------------------------------------------------------
#> ind60 =~ x1 |        0.92 | 0.02 | [0.87, 0.96] | < .001
#> ind60 =~ x2 |        0.97 | 0.02 | [0.94, 1.01] | < .001
#> ind60 =~ x3 |        0.87 | 0.03 | [0.81, 0.93] | < .001
#> dem60 =~ y1 |        0.85 | 0.04 | [0.77, 0.93] | < .001
#> dem60 =~ y2 |        0.72 | 0.06 | [0.59, 0.84] | < .001
#> dem60 =~ y3 |        0.72 | 0.06 | [0.60, 0.85] | < .001
#> dem60 =~ y4 |        0.85 | 0.04 | [0.76, 0.93] | < .001
#> dem65 =~ y5 |        0.81 | 0.05 | [0.71, 0.90] | < .001
#> dem65 =~ y6 |        0.75 | 0.06 | [0.63, 0.86] | < .001
#> dem65 =~ y7 |        0.82 | 0.04 | [0.74, 0.91] | < .001
#> dem65 =~ y8 |        0.83 | 0.04 | [0.74, 0.92] | < .001
#>
#> # Regression
#>
#> Link          | Coefficient |   SE |       95% CI |      p
#> ----------------------------------------------------------
#> dem60 ~ ind60 |        0.45 | 0.10 | [0.24, 0.65] | < .001
#> dem65 ~ ind60 |        0.18 | 0.07 | [0.04, 0.32] | 0.010
#> dem65 ~ dem60 |        0.89 | 0.05 | [0.79, 0.98] | < .001
#>
#> # Correlation
#>
#> Link     | Coefficient |   SE |        95% CI |      p
#> ------------------------------------------------------
#> y1 ~~ y5 |        0.30 | 0.14 | [ 0.02, 0.57] | 0.034
#> y2 ~~ y4 |        0.27 | 0.12 | [ 0.04, 0.51] | 0.024
#> y2 ~~ y6 |        0.36 | 0.10 | [ 0.17, 0.55] | < .001
#> y3 ~~ y7 |        0.19 | 0.13 | [-0.07, 0.45] | 0.152
#> y4 ~~ y8 |        0.11 | 0.13 | [-0.15, 0.37] | 0.405
#> y6 ~~ y8 |        0.34 | 0.11 | [ 0.12, 0.55] | 0.002
```
