### 2020-08-26 easystats/parameters#290 @strengejacke
I think this can be closed, the function now also returns self-defined parameters.
