### 2020-08-27 #2445 @Ohlsen
I finished my review, too. I had some feature requests focusing in particular on the packages functionality towards lavaan-models, which are already solved . I recommend accept as well.
