### 2020-08-27 #2445 @strengejacke
Thanks, @Ohlsen, for your suggestions. Regarding your issues:

1) We revised the output for structural equation models, so `model_parameters()` now also include self-defined parameters (https://github.com/easystats/parameters/issues/290).
2) Since the vignettes on `model_parameters()` for the _lavaan_ package only focussed on CFA, we also added a small example for SEM as well.
3) We fixed some typos found in the paper.
4) To improve support for structural equation models, we revised `model_parameters()` so it can include different types of standardized solutions (https://github.com/easystats/parameters/issues/291).

We hope we adequately addressed all issues!
