+++ R/p_value.R
@@ -1489,7 +1489,7 @@ p_value.lavaan <- function(model, ...) {
 
 #' @export
 p_value.blavaan <- function(model, ci = .95, ...) {
-  out <- .extract_parameters_blavaan(model, ...)
+  out <- .extract_parameters_lavaan(model, ...)
   out[out$Operator != "~1", c("To", "Operator", "From", "p")]
 }
 
