+++ tests/testthat/test-model_parameters_std_mixed.R
@@ -40,12 +40,14 @@ if (.runThisTest) {
       testthat::expect_equal(params$CI_high, c(0, 0.84893, 0.41735, 0.46488, 2.01523, -0.19075, -0.01014), tolerance = 1e-3)
     })
 
-    test_that("model_parameters, standardize-refit robust", {
-      params <- model_parameters(model, standardize = "refit", robust = TRUE, vcov_estimation = "CR", vcov_type = "CR1", vcov_args = list(cluster = iris$grp))
-      testthat::expect_equal(c(nrow(params), ncol(params)), c(7, 8))
-      testthat::expect_equal(params$Coefficient, c(0.96949, -1.28631, -1.81461, 0.34791, 1.74252, -0.25421, -0.18834), tolerance = 1e-3)
-      testthat::expect_equal(params$SE, c(0.07726, 0.33406, 0.22647, 0.0524, 0.10092, 0.18537, 0.05552), tolerance = 1e-3)
-      testthat::expect_equal(params$CI_high, c(1.12224, -0.6259, -1.36691, 0.45151, 1.94204, 0.11227, -0.07858), tolerance = 1e-3)
-    })
+    if (require("clubSandwich")) {
+      test_that("model_parameters, standardize-refit robust", {
+        params <- model_parameters(model, standardize = "refit", robust = TRUE, vcov_estimation = "CR", vcov_type = "CR1", vcov_args = list(cluster = iris$grp))
+        testthat::expect_equal(c(nrow(params), ncol(params)), c(7, 8))
+        testthat::expect_equal(params$Coefficient, c(0.96949, -1.28631, -1.81461, 0.34791, 1.74252, -0.25421, -0.18834), tolerance = 1e-3)
+        testthat::expect_equal(params$SE, c(0.07726, 0.33406, 0.22647, 0.0524, 0.10092, 0.18537, 0.05552), tolerance = 1e-3)
+        testthat::expect_equal(params$CI_high, c(1.12224, -0.6259, -1.36691, 0.45151, 1.94204, 0.11227, -0.07858), tolerance = 1e-3)
+      })
+    }
   }
 }
