### 2020-08-27 #2445 @sbenthall
Thank you, reviewers.

@strengejacke , as a next step, would you please :
* issue a new tagged release of the software (if changed),
* archive it (on Zenodo, figshare, or other).
* post the version number and archive DOI in this REVIEW issue.
