+++ NEWS.md
@@ -1,4 +1,4 @@
-# parameters 0.8.2.1
+# parameters 0.8.3
 
 ## General
 
