### 2020-08-27 #2445 @strengejacke
Thank you @sbenthall

- new release with latest changes from the revisions of this review is tagged: https://github.com/easystats/parameters/releases/tag/0.8.3
- archive for latest release on Zenodo created: https://zenodo.org/record/4004846#.X0gIBMgzaUk
- Version number: 0.8.3; DOI: 10.5281/zenodo.4004846
