+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 2.7.3
 pkgdown: 1.5.1.9000
-pkgdown_sha: eae56f08694abebf93cdfc0dd8e9ede06d8c815f
+pkgdown_sha: ecce89ed7fc0bfe4a2ec6e2fdc9285f238e13e7c
 articles: []
-last_built: 2020-08-03T13:35Z
+last_built: 2020-08-27T19:48Z
 
