+++ pkgdown.yml
@@ -2,5 +2,5 @@ pandoc: 2.7.3
 pkgdown: 1.5.1.9000
 pkgdown_sha: ecce89ed7fc0bfe4a2ec6e2fdc9285f238e13e7c
 articles: []
-last_built: 2020-08-27T19:48Z
+last_built: 2020-08-30T10:41Z
 
