+++ NAMESPACE
@@ -778,6 +778,7 @@ importFrom(insight,get_varcov)
 importFrom(insight,get_variance)
 importFrom(insight,has_intercept)
 importFrom(insight,is_model)
+importFrom(insight,is_multivariate)
 importFrom(insight,model_info)
 importFrom(insight,n_obs)
 importFrom(insight,print_color)
