### 2020-08-31 7efcedf
catch test exceptions
