+++ NAMESPACE
@@ -746,6 +746,7 @@ importFrom(bayestestR,convert_pd_to_p)
 importFrom(bayestestR,describe_posterior)
 importFrom(bayestestR,equivalence_test)
 importFrom(bayestestR,p_direction)
+importFrom(bayestestR,reshape_ci)
 importFrom(bayestestR,rope_range)
 importFrom(grDevices,colorRampPalette)
 importFrom(grDevices,dev.off)
