### 2020-09-09 #2445 @danielskatz
> Your submission is accepted, @strengejacke
>
> Thank you for your article!

@sbenthall - FYI, this is not accepted, it's "recommended for acceptance".  See the label...

Accepted happens when it is published, which I hope will happen soon :)
