### 2020-09-09 #2445 @danielskatz
Once these are done, I will finalize the acceptance and publication
